requirements.txt
aiohttp
python-dotenv
google-genai